feat: adiciona repositorio e Dao de Pedido

+++ src/main/java/com/lucas/sistema/entrega/modules/pedido/application/dto/PedidoAdicionarRequest.java
@@ -4,5 +4,5 @@ import com.lucas.sistema.entrega.modules.pedido.domain.enumerator.PedidoStatus;
 
 import java.time.LocalDateTime;
 
-public record PedidoAdicionarRequest(long clienteId, LocalDateTime dataPedido, int volumeM3, int pesoKg, PedidoStatus pedidoStatus) {
+public record PedidoAdicionarRequest(long clienteId, LocalDateTime dataPedido, double volumeM3, double pesoKg, PedidoStatus pedidoStatus) {
 }
